Extract `SentryConfig` struct
