+++ src/config.rs
@@ -1,10 +1,12 @@
 mod balance_capacity;
 mod base;
 mod database_pools;
+mod sentry;
 mod server;
 
 pub use self::balance_capacity::BalanceCapacityConfig;
 pub use self::base::Base;
 pub use self::database_pools::{DatabasePools, DbPoolConfig};
+pub use self::sentry::SentryConfig;
 pub(crate) use self::server::domain_name;
 pub use self::server::Server;
